crawler2/frontier: Add thread-safe Frontier class
The second step towards a muli-threaded crawler. Additionally, this commit
changes the old frontier interface to accept nurls (node URLS) instead of URL
strings. To reflect this, the method names now include "nurl" instead of "url".

This does cause incompatability with the existing codebase, but it shouldn't be
difficult to adjust.