scraper2: Add process_text, change scraper return value
This adds the ability to process text. Currently, this only computes the word
counts, tokens, and content size to be consumed by the scraper function.

Scraper return value changed to a result tuple (ok, err). This helps communicate
to the crawler what to do in case the scraper fails to scrape any nurls.