crawler2/nurl: Expand status to int
This adds an intermediate state to status. The intermediate state helps avoid
concurrency issues when retrieving from the nurls deque. The frontier shall
assign an intermediate state iff the thread has consumed the nurl but is not yet
done.

This should fix the duplicate checking of nurls, single-threaded and
multi-threaded.