crawler2/frontier: Fix loading from intermediate states
This updates URLs that are in the intermediate state (status=0x1) to the
un-downloaded state upon nap initialization. This fixes the frontier not
downloading from its pre-existing set of URLs.

Additionally, change "add_nurl" to add nurl to nap iff it doesn't exist, which
guarantees any nurls added to the deque shall exist. This change also fixes a
race condition in which a thread marks a nurl as complete but its derived nurls
have not updated to the nap. At worst, the crawler may re-download the nurl
again if the derived nurls were added to nap but the nurl was not marked
complete. This is rather minimal compared to the possible loss of nurls prior to
the change. In short, concurrency hell.