crawler2/nurl: Add "hash", remove "exlinks", "smlinks", and "size"
Hash added to speed up the information retrieval without the need to re-compute
the hash or import the necessary utils.

Exlinks, smlinks, and size were removed since it complicates the retrieval of
matching similar/exact pages. The intent is to add it to the Nap object itself
so the worker can process it after scraping.