helpers/simhash: Use crc64 for hashing
This guarantees that a hash of the same string will remain unchanged even if a
new Python process is launched. Note that Python's built-in hash() function
effectively randomizes the hash for security purposes, which is not ideal for
our purposes.

The crc64 is modified from "helpers/crc32.py" with a new 64-bit polynomial based
on CRC-64/XZ
(https://reveng.sourceforge.io/crc-catalogue/all.htm#crc.cat.crc-64-xz).

The polynomial has been tested against a few websites and seems to generate
"random" hashes.