crawler2/worker: Convert worker thread to a pipeline
The previous worker thread seems hacked together to support the necessary
features. This refactors the thread, so it's easier to read and change behavior
in-between pipelines.

This also tells the crawler to download using requests.get instead of through
the cache server to test the crawler easily. This should be changed later!