crawler2/frontier: Lock/unlock immediately after downloading robots.txt
Note that it's possible for a URL to break the crawl-delay once it downloads
robot.txt if it never forces the domain specific polite mutex to lock/unlock.
Consider a crawl-delay of 10 seconds: a crawler might download after 0.5 seconds
(the global politeness timer), which violates the crawl-delay.

This change makes it so that crawl-delay shall be respected even after
downloading robots.txt.