scraper2: Refactor sitemap parsing
This refactors the sitemap parser to use the lxml-xml parser for bs4.

It also cleans up how sitemaps are parsed by introducing separate functions for
the <sitemapindex> tag and the <urlset> tag, which are required by the Sitemaps
protocol (https://www.sitemaps.org/protcol.html).

ParsedResponse now has an additional attribute to mark it as a sitemap in case
the worker chooses to filter by text_content.