crawler2/frontier: Fix parent assignment for sitemaps
This manually sets the attributes instead of using Nurl.set_parent() since the
robots URL does not exist as a nurl. The parent is stored unhashed as the robots
URL.

To allow for this change, "crawler2/nurl.py" was modified to document that the
parent may include unhashed URLs only if the parent URL doesn't exist as a Nurl
object.