crawler2/frontier: Remove traversal policy
This change is more about choosing a simpler approach when getting nurls in a
multi-threaded way.

The previous method, which used deques as the underlying implementation, did not
support tasks, which means if a deque was empty, the thread would kill itself
even if a URL was still being processed. This was a bug that would've required
extending deque to support tasks.

I decided against this and replaced "deque" with "Queue" instead to reduce the
work required to fix this issue. Unfortunately, this requires scrapping
traversal policies altogether.

It's possible to add back traversal policy if deques with tasks were to be
implemented, but I don't think it's worth adding yet.

Note: The workers for crawler2 and crawlerman have been modified to support
Queues by using task_done() which indicates when a "get" request has completely
processed.