crawler2/worker: Move pipe functions to separate file
This moves the worker pipe functions to "crawler2/workerpipe.py". The rationale
behind this change is make the worker thread itself not be bloated with all
declarations of the pipe functions. Instead, hooking into the pipeline is as
simple as importing workerpipe.

This also changes crawlerman/worker.py to import crawler2/workerpipe instead.